abn Package

abn is a collection of functions for fitting, selecting/learning, analyzing, reporting Additive Bayesian Networks.

General overview

What is abn: Bayesian network modeling is a data analysis technique that is ideally suited to messy, highly correlated, and complex data sets. This methodology is somewhat distinct from other forms of statistical modeling in that its focus is on structure discovery - determining an optimal graphical model that describes the inter-relationships in the underlying processes which generated the data. It is a multivariate technique and can be used for one or many dependent variables. This is a data-driven approach, as opposed to, rely only on subjective expert opinion to determine how variables of interest are inter-related (for example, structural equation modeling).

The R package abn is designed to fit additive Bayesian models to observational data sets. It contains routines to score Bayesian Networks based on Bayesian or information-theoretic formulation of generalized linear models. It is equipped with exact search and greedy search algorithms to select the best network. The Bayesian implementation supports random effects to control for one layer clustering. It supports a possible mixture of continuous, discrete, and count data and inputs of prior knowledge at a structural level.

The R package abn requires the R package Rgraphviz to work well. It is store outside of CRAN; see ‘Examples’ for the code to install the last version.

The web page r-bayesian-networks.org provides further case studies. See also the files provided in the package directories inst/bootstrapping_example and inst/old_vignette for more details.

References

Kratzer, Gilles, Fraser Lewis, Arianna Comin, Marta Pittavino, and Reinhard Furrer. “Additive Bayesian Network Modeling with the R Package Abn.” Journal of Statistical Software 105 (January 28, 2023): 1–41. https://doi.org/10.18637/jss.v105.i08.

Kratzer, G., Lewis, F.I., Comin, A., Pittavino, M. and Furrer, R. (2019). "Additive Bayesian Network Modelling with the R Package abn". arXiv preprint arXiv:1911.09006.

Lewis, F. I., and Ward, M. P. (2013). "Improving epidemiologic data analyses through multivariate regression modeling". Emerging themes in epidemiology, 10(1), 4.

Kratzer, G., Pittavino, M, Lewis, F. I., and Furrer, R., (2017). "abn: an R package for modelling multivariate data using additive Bayesian networks". R package version 2.2. https://CRAN.R-project.org/package=abn

See also

Author

Maintainer: Matteo Delucchi matteo.delucchi@math.uzh.ch (ORCID)

Authors:

Reinhard Furrer reinhard.furrer@math.uzh.ch (ORCID)

Gilles Kratzer gilles.kratzer@gmail.com (ORCID)

Fraser Iain Lewis fraser.iain.lewis@gmail.com (ORCID)

Other contributors:

Marta Pittavino marta.pittavino@math.uzh.ch (ORCID) [contributor]

Kalina Cherneva kalinacherneva@gmail.com [contributor]

Examples

## Citations:
print(citation('abn'), bibtex=TRUE)
#> To cite the methodology of the R package 'abn' use:
#> 
#>   Kratzer G, Lewis F, Comin A, Pittavino M, Furrer R (2023). “Additive
#>   Bayesian Network Modeling with the R Package abn.” _Journal of
#>   Statistical Software_, *105*(8), 1-41. doi:10.18637/jss.v105.i08
#>   <https://doi.org/10.18637/jss.v105.i08>.
#> 
#> A BibTeX entry for LaTeX users is
#> 
#>   @Article{,
#>     title = {Additive Bayesian Network Modeling with the R Package abn},
#>     author = {Gilles Kratzer and Fraser Lewis and Arianna Comin and Marta Pittavino and Reinhard Furrer},
#>     journal = {Journal of Statistical Software},
#>     year = {2023},
#>     volume = {105},
#>     number = {8},
#>     pages = {1--41},
#>     doi = {10.18637/jss.v105.i08},
#>   }
#> 
#> To cite an example of a typical ABN analysis use:
#> 
#>   Kratzer, G., Lewis, F.I., Willi, B., Meli, M.L., Boretti, F.S.,
#>   Hofmann-Lehmann, R., Torgerson, P., Furrer, R. and Hartnack, S.
#>   (2020). Bayesian Network Modeling Applied to Feline Calicivirus
#>   Infection Among Cats in Switzerland. Frontiers in Veterinary Science,
#>   7, 73
#> 
#> A BibTeX entry for LaTeX users is
#> 
#>   @Article{,
#>     title = {Bayesian Network Modeling Applied to Feline Calicivirus Infection Among Cats in Switzerland},
#>     author = {Gilles Kratzer and Fraser Iain Lewis and Barbara Willi and Marina Meli and Felicitas Boretti and Regina Hofmann-Lehmann and Paul Torgerson and Reinhard Furrer and Sonja Hartnack},
#>     year = {2020},
#>     journal = {Frontiers in Veterinary Science},
#>   }
#> 
#> To cite the software implementation of the R package 'abn' use:
#> 
#>   Furrer, R., Kratzer, G. and Lewis, F.I. (2023). abn: Modelling
#>   Multivariate Data with Additive Bayesian Networks. R package version
#>   2.7-2.  https://CRAN.R-project.org/package=abn
#> 
#> A BibTeX entry for LaTeX users is
#> 
#>   @Manual{,
#>     title = {abn: Modelling Multivariate Data with Additive Bayesian Networks},
#>     author = {Reinhard Furrer and Gilles Kratzer and Fraser Iain Lewis},
#>     year = {2023},
#>     note = {R package version 2.7-2},
#>     url = {https://CRAN.R-project.org/package=abn},
#>   }

## Installing the R package Rgraphviz:
# if (!requireNamespace("BiocManager", quietly = TRUE))
#     install.packages("BiocManager")
# BiocManager::install("Rgraphviz")

## README.md in the directory `bootstrapping_example/`:
# edit(file=paste0( path.package('abn'),'/bootstrapping_example/README.md'))