Fit a given regression using INLA

Internal wrapper to INLA and are called from fitAbn.bayes and buildScoreCache.bayes.

Usage

calc.node.inla.glmm(
  child.loc = NULL,
  dag.m.loc = NULL,
  data.df.loc = NULL,
  data.dists.loc = NULL,
  ntrials.loc = NULL,
  exposure.loc = NULL,
  compute.fixed.loc = NULL,
  mean.intercept.loc = NULL,
  prec.intercept.loc = NULL,
  mean.loc = NULL,
  prec.loc = NULL,
  loggam.shape.loc = NULL,
  loggam.inv.scale.loc = NULL,
  verbose.loc = FALSE,
  nthreads = NULL
)

Arguments

child.loc

index of current child node.

dag.m.loc

dag as matrix.

data.df.loc

data df,

data.dists.loc

list of distributions.

ntrials.loc

rep(1,dim(data.df)[1]).

exposure.loc

rep(1,dim(data.df)[1]).

compute.fixed.loc

TRUE.

mean.intercept.loc

the prior mean for all the Gaussian additive terms for each node. INLA argument control.fixed=list(mean.intercept=...) and control.fixed=list(mean=...).

prec.intercept.loc

the prior precision for all the Gaussian additive term for each node. INLA argument control.fixed=list(prec.intercept=...) and control.fixed=list(prec=...).

mean.loc

the prior mean for all the Gaussian additive terms for each node. INLA argument control.fixed=list(mean.intercept=...) and control.fixed=list(mean=...). Same as mean.intercept.loc.

prec.loc

the prior precision for all the Gaussian additive term for each node. INLA argument control.fixed=list(prec.intercept=...) and control.fixed=list(prec=...). Same as prec.intercept.loc.

loggam.shape.loc

the shape parameter in the Gamma distribution prior for the precision in a Gaussian node. INLA argument control.family=list(hyper = list(prec = list(prior="loggamma",param=c(loggam.shape, loggam.inv.scale)))).

loggam.inv.scale.loc

the inverse scale parameter in the Gamma distribution prior for the precision in a Gaussian node. INLA argument control.family=list(hyper = list(prec = list(prior="loggamma",param=c(loggam.shape, loggam.inv.scale)))).

verbose.loc

FALSE to not print additional output.

nthreads

number of threads to use for INLA. Default is fit.control[["ncores"]] or build.control[["ncores"]] which is the number of cores specified in control and defaults to 1.

Value

If INLA failed, FALSE or an error is returned. Otherwise, the direct output from INLA is returned.

See also